Move run command implementation out of app
